Mask propagation is used for shape pruning
