+++ nncf/torch/pruning/filter_pruning/algo.py
@@ -33,8 +33,8 @@ from nncf.common.pruning.statistics import FilterPruningStatistics
 from nncf.common.pruning.statistics import PrunedModelTheoreticalBorderline
 from nncf.common.pruning.statistics import PrunedLayerSummary
 from nncf.common.pruning.statistics import PrunedModelStatistics
-from nncf.common.pruning.utils import ShapePruninigProcessor
-from nncf.common.pruning.utils import WeightsFlopsCalculator
+from nncf.common.pruning.shape_pruning import ShapePruninigProcessor
+from nncf.common.pruning.shape_pruning import WeightsFlopsCalculator
 from nncf.common.pruning.utils import get_rounded_pruned_element_number
 from nncf.common.schedulers import StubCompressionScheduler
 from nncf.common.statistics import NNCFStatistics
@@ -136,6 +136,7 @@ class FilterPruningController(BasePruningAlgoController):
                                                           conv_op_metatypes=GENERAL_CONV_LAYER_METATYPES,
                                                           linear_op_metatypes=LINEAR_LAYER_METATYPES)
         self._shape_pruning_proc = ShapePruninigProcessor(graph=graph,
+                                                          pruning_operations_metatype=PT_PRUNING_OPERATOR_METATYPES,
                                                           pruning_groups=pruned_module_groups,
                                                           prunable_types=prunable_types)
         self._init_module_channels_and_shapes()
